SOCINT-338 Refactoring as per 338 ticket
